[CDS-41062]: small refactoring
